feat(manuscript/date): добавить создания дат для года (связь movable_day & day), к manuscript добавить Bookmark - это meny to meny (manuscript to book)
Так же протестировать создания dates, но т.к. тесты долго выполняются, не вызывать их при вызове всех тестов

К manuscript добавить Bookmark - закладки, связываясь с book
Bookmark - meny to meny (manuscript to book)

Добавить возможность добавлять bookmark к manuscript через api
Так же добавить возможность обновления manuscript через api

feat(manuscript) - ЕЩЕ НЕ ПРОТЕСТИРОВАНО

---

Стоит быть очень аккуратным со связями pydantic schemas
Писать в
    if TYPE_CHECKING:
только те импорты, которые вызовут ошибку ImportError, если их там не написать
Вместе с TYPE_CHECKING добавлять и
from __future__ import annotations
(т.к без этого будет ошибка)
Если не писать TYPE_CHECKING, то и from __future__ import annotations так же не нужно писать

Заметил, что при большом кол. связей схем, если на пути связей встречается в какой-то из моделей связь с TYPE_CHECKING, то ВЫДАЕТ ОШИБКУ
т.е если изначально была связь без TYPE_CHECKING, то и везде дальше так же все схемы должны быть связаны без импортов с TYPE_CHECKING
(но это просто из наблюдений/ручных тестов, может быть ошибки были и в другом)

Так же создавать много pydantic схем, чтобы схема при связях не начала ссылаться на саму себя (и получилась рекурсия связей)

---

Так же создать класс PrepareYearTitle, но еще не добавлять/связывать его с YearCreate
PrepareYearTitle -будет пока что использоваться для prepare year с libs (rsl, nlr, neb)

+++ app/tests/test_utils/create_test_data.py
@@ -26,56 +26,68 @@ def create_dirs_for_test_data():
             sub_dir_path.mkdir(exist_ok=True)
 
 
-def create_saints_data(session: requests.Session):
+def create_get_saints_data(session: requests.Session):
     path = Path(settings.TEST_DATA_DIR) / 'saint/some_saint_slug.html'
     for saint_slug in [
         'ioann-zlatoust',
         'marija-egipetskaja',
         'igor-v-kreshchenii-georgij-chernigovskij-i-kievskij'
     ]:
-        saint_data: Tag = collect_saint_data(session, saint_slug=saint_slug)
         current_path = path.with_stem(saint_slug)
-        current_path.write_text(str(saint_data), encoding="utf-8")
+        if not current_path.exists():
+            saint_data: Tag = collect_saint_data(session, saint_slug=saint_slug)
+            current_path.write_text(str(saint_data), encoding="utf-8")
 
 
-def create_manuscripts_data_rsl(session: requests.Session):
+def create_get_manuscripts_data_rsl(session: requests.Session):
     path = Path(settings.TEST_DATA_DIR) / 'manuscript/get/rsl/some_manuscript_code.html'
     for manuscript_code in [
         'f-304iii-15',
         'f-173i-2',
         'f-98-36',
-        'f-98-80'
+        'f-98-80',
+        'f-7-16',
+        'f-37-15'
     ]:
-        collect_manuscript_data = CollectManuscriptDataFromRsl(session, manuscript_code=manuscript_code)
-        manuscript_data: Tag = collect_manuscript_data._soup
         current_path = path.with_stem(manuscript_code)
-        current_path.write_text(str(manuscript_data), encoding="utf-8")
+        if not current_path.exists():
+            collect_manuscript_data = CollectManuscriptDataFromRsl(session, manuscript_code=manuscript_code)
+            manuscript_data: Tag = collect_manuscript_data._soup
+            current_path.write_text(str(manuscript_data), encoding="utf-8")
 
 
-def create_manuscripts_data_neb(session: requests.Session):
+def create_get_manuscripts_data_neb(session: requests.Session):
     path = Path(settings.TEST_DATA_DIR) / 'manuscript/get/neb/some_manuscript_neb_slug.html'
     for manuscript_neb_slug in [
         'apostol-aprakos-licevaya-rukopis',
         'prolog-mart-avgust-4',
-        'evangelie-uchitelnoe-licevoe'
+        'evangelie-uchitelnoe-licevoe',
+        'psaltyr-s-vossledovaniem-1',
+        'sbornik-slov-kanonov-i-zhitiy',
+        'zhitie-nikolaya-chudotvorca-licevoe-otryvok-iz-letopisnogo-svoda-licevoy'
     ]:
-        collect_manuscript_data = CollectManuscriptDataFromNeb(session, manuscript_code=manuscript_neb_slug)
-        manuscript_data: Tag = collect_manuscript_data._soup
         current_path = path.with_stem(manuscript_neb_slug)
-        current_path.write_text(str(manuscript_data), encoding="utf-8")
+        if not current_path.exists():
+            collect_manuscript_data = CollectManuscriptDataFromNeb(session, manuscript_code=manuscript_neb_slug)
+            manuscript_data: Tag = collect_manuscript_data._soup
+            current_path.write_text(str(manuscript_data), encoding="utf-8")
 
 
-def create_manuscripts_data_nlr(session: requests.Session):
+def create_get_manuscripts_data_nlr(session: requests.Session):
     path = Path(settings.TEST_DATA_DIR) / 'manuscript/get/nlr/some_manuscript_code.json'
     for manuscript_code in [
         '14CB689B-8344-49F0-9DD6-0AAE09460BE0',
         '59014138-46CC-4335-9E3C-9F903FD854A3',
-        '919048C3-8920-40C1-A94D-AA216B5E2FD6'
+        '919048C3-8920-40C1-A94D-AA216B5E2FD6',
+        'D157FA07-FF83-4AF2-851A-09DC187B4F9F',
+        '2EADD685-EB71-40C5-BDEC-7E9DC9B7EFCE',
+        '466FB16F-C1B8-443E-BD95-C02D272D0251'
     ]:
-        r = session.post(const.NlrUrl.GET_MANUSCRIPT_DATA_API, data={'ab': manuscript_code})
-        manuscript_data_json: str = r.text
         current_path = path.with_stem(manuscript_code)
-        current_path.write_text(manuscript_data_json, encoding="utf-8")
+        if not current_path.exists():
+            r = session.post(const.NlrUrl.GET_MANUSCRIPT_DATA_API, data={'ab': manuscript_code})
+            manuscript_data_json: str = r.text
+            current_path.write_text(manuscript_data_json, encoding="utf-8")
 
 
 def create_search_manuscripts_nlr(session: requests.Session):
@@ -85,14 +97,15 @@ def create_search_manuscripts_nlr(session: requests.Session):
         ('Кир.-Бел. 1/1240', '59014138-46CC-4335-9E3C-9F903FD854A3'),
         ('F.IV.225', '919048C3-8920-40C1-A94D-AA216B5E2FD6')
     ]:
-        search_data: dict[str, str] = {
-            'doc_Shifr': manuscript_code_title,
-            'doc_IsYesDigital': 'on'
-        }
-        r = session.post(const.NlrUrl.SEARCH_MANUSCRIPT_API, data=search_data)
-        search_manuscript_json: str = r.text
         current_path = path.with_stem(manuscript_code)
-        current_path.write_text(search_manuscript_json, encoding="utf-8")
+        if not current_path.exists():
+            search_data: dict[str, str] = {
+                'doc_Shifr': manuscript_code_title,
+                'doc_IsYesDigital': 'on'
+            }
+            r = session.post(const.NlrUrl.SEARCH_MANUSCRIPT_API, data=search_data)
+            search_manuscript_json: str = r.text
+            current_path.write_text(search_manuscript_json, encoding="utf-8")
 
 
 def create_search_manuscripts_neb(session: requests.Session):
@@ -101,25 +114,29 @@ def create_search_manuscripts_neb(session: requests.Session):
         ('F.I.60', 'apostol-aprakos-licevaya-rukopis'),
         ('Соф. 1345', 'prolog-mart-avgust-4'),
         ('Ф.98 №80', 'evangelie-uchitelnoe-licevoe'),
+        ('Сол. 754/864', 'psaltyr-s-vossledovaniem-1'),
+        ('Кир.-Бел. 19/1096', 'sbornik-slov-kanonov-i-zhitiy'),
+        ('Ф.37 №15', 'zhitie-nikolaya-chudotvorca-licevoe-otryvok-iz-letopisnogo-svoda-licevoy'),
         ('NOT_FOUND_IN_NEB', 'NOT_FOUND_IN_NEB')
     ]:
-        url: str = SearchManuscriptInNeb.prepare_neb_search_manuscript_api_url(manuscript_code_title)
-        r = session.get(url)
-        search_manuscript_json: str = r.text
         current_path = path.with_stem(manuscript_neb_slug)
-        current_path.write_text(search_manuscript_json, encoding="utf-8")
+        if not current_path.exists():
+            url: str = SearchManuscriptInNeb.prepare_neb_search_manuscript_api_url(manuscript_code_title)
+            r = session.get(url)
+            search_manuscript_json: str = r.text
+            current_path.write_text(search_manuscript_json, encoding="utf-8")
 
 
-def create_all_test_data():
+def create_all_test_data(session: requests.Session):
     # create_dirs_for_test_data()
-    session: requests.Session = next(deps.get_session())
-    # create_saints_data(session)
-    # create_manuscripts_data_rsl(session)
-    # create_manuscripts_data_neb(session)
-    # create_manuscripts_data_nlr(session)
-    # create_search_manuscripts_nlr(session)
+    create_get_saints_data(session)
+    create_get_manuscripts_data_rsl(session)
+    create_get_manuscripts_data_neb(session)
+    create_get_manuscripts_data_nlr(session)
+    create_search_manuscripts_nlr(session)
     create_search_manuscripts_neb(session)
 
 
 if __name__ == '__main__':
-    create_all_test_data()
+    session: requests.Session = next(deps.get_session())
+    create_all_test_data(session)
